--- src/test/java/no/difi/sdp/client2/smoke/ObjectMother.java
@@ -1,113 +1,0 @@
-package no.difi.sdp.client2.smoke;
-
-import no.difi.sdp.client2.domain.Avsender;
-import no.difi.sdp.client2.domain.Dokument;
-import no.difi.sdp.client2.domain.Dokumentpakke;
-import no.difi.sdp.client2.domain.Forsendelse;
-import no.difi.sdp.client2.domain.Mottaker;
-import no.difi.sdp.client2.domain.Noekkelpar;
-import no.difi.sdp.client2.domain.Prioritet;
-import no.difi.sdp.client2.domain.Sertifikat;
-import no.difi.sdp.client2.domain.Databehandler;
-import no.difi.sdp.client2.domain.digital_post.DigitalPost;
-import no.difi.sdp.client2.domain.digital_post.EpostVarsel;
-import no.difi.sdp.client2.domain.digital_post.Sikkerhetsnivaa;
-import no.difi.sdp.client2.domain.digital_post.SmsVarsel;
-import no.digipost.api.representations.Organisasjonsnummer;
-
-import java.io.InputStream;
-import java.util.ArrayList;
-import java.util.Date;
-import java.util.UUID;
-
-import static java.util.Arrays.asList;
-
-public class ObjectMother {
-
-    public static Forsendelse forsendelse(String orgNumber, String mpcId, InputStream dokumentStream) {
-        DigitalPost digitalPost = digitalPost();
-
-        Dokument hovedDokument = Dokument.builder("Sensitiv brevtittel", "faktura.pdf", dokumentStream)
-                .mimeType("application/pdf")
-                .build();
-
-        Dokumentpakke dokumentpakke = Dokumentpakke.builder(hovedDokument)
-                .vedlegg(new ArrayList<Dokument>())
-                .build();
-
-        Avsender avsender = avsender(orgNumber);
-
-        return Forsendelse.digital(avsender, digitalPost, dokumentpakke)
-                .konversasjonsId(UUID.randomUUID().toString())
-                .prioritet(Prioritet.PRIORITERT)
-                .mpcId(mpcId)
-                .spraakkode("NO")
-                .build();
-    }
-
-    public static Databehandler databehandlerMedSertifikat(final Organisasjonsnummer organisasjonsnummer, final Noekkelpar noekkelpar) {
-        return Databehandler.builder(organisasjonsnummer, noekkelpar)
-                .build();
-    }
-
-    public static DigitalPost digitalPost() {
-        EpostVarsel epostVarsel = EpostVarsel.builder("example@email.org", "Du har mottatt brev i din digitale postkasse")
-                .varselEtterDager(asList(1, 4, 10))
-                .build();
-
-        Mottaker mottaker = Mottaker.builder("04036125433", "ove.jonsen#6K5A", mottakerSertifikat(), Organisasjonsnummer.of("984661185"))
-                .build();
-
-        SmsVarsel smsVarsel = SmsVarsel.builder("4799999999", "Du har mottatt brev i din digitale postkasse")
-                .build();
-
-        return DigitalPost.builder(mottaker, "Ikke-sensitiv tittel for forsendelsen")
-                .virkningsdato(new Date())
-                .aapningskvittering(false)
-                .sikkerhetsnivaa(Sikkerhetsnivaa.NIVAA_3)
-                .epostVarsel(epostVarsel)
-                .smsVarsel(smsVarsel)
-                .build();
-    }
-
-
-    public static Avsender avsender(final String orgNumber) {
-        return Avsender.builder(orgNumber)
-                .build();
-    }
-
-    public static Sertifikat mottakerSertifikat() {
-        return DigipostMottakerSertifikatTest();
-    }
-
-    private static Sertifikat DigipostMottakerSertifikatTest() {
-        return Sertifikat.fraBase64X509String(
-                "MIIE7jCCA9agAwIBAgIKGBZrmEgzTHzeJjANBgkqhkiG9w0BAQsFADBRMQswCQYD" +
-                "VQQGEwJOTzEdMBsGA1UECgwUQnV5cGFzcyBBUy05ODMxNjMzMjcxIzAhBgNVBAMM" +
-                "GkJ1eXBhc3MgQ2xhc3MgMyBUZXN0NCBDQSAzMB4XDTE0MDQyNDEyMzA1MVoXDTE3" +
-                "MDQyNDIxNTkwMFowVTELMAkGA1UEBhMCTk8xGDAWBgNVBAoMD1BPU1RFTiBOT1JH" +
-                "RSBBUzEYMBYGA1UEAwwPUE9TVEVOIE5PUkdFIEFTMRIwEAYDVQQFEwk5ODQ2NjEx" +
-                "ODUwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQCLCxU4oBhtGmJxXZWb" +
-                "dWdzO2uA3eRNW/kPdddL1HYl1iXLV/g+H2Q0ELadWLggkS+1kOd8/jKxEN++biMm" +
-                "mDqqCWbzNdmEd1j4lctSlH6M7tt0ywmXIYdZMz5kxcLAMNXsaqnPdikI9uPJZQEL" +
-                "3Kc8hXhXISvpzP7gYOvKHg41uCxu1xCZQOM6pTlNbxemBYqvES4fRh2xvB9aMjwk" +
-                "B4Nz8jrIsyoPI89i05OmGMkI5BPZt8NTa40Yf3yU+SQECW0GWalB5cxaTMeB01tq" +
-                "slUzBJPV3cQx+AhtQG4hkOhQnAMDJramSPVtwbEnqOjQ+lyNmg5GQ4FJO02ApKJT" +
-                "ZDTHAgMBAAGjggHCMIIBvjAJBgNVHRMEAjAAMB8GA1UdIwQYMBaAFD+u9XgLkqNw" +
-                "IDVfWvr3JKBSAfBBMB0GA1UdDgQWBBQ1gsJfVC7KYGiWVLP7ZwzppyVYTTAOBgNV" +
-                "HQ8BAf8EBAMCBLAwFgYDVR0gBA8wDTALBglghEIBGgEAAwIwgbsGA1UdHwSBszCB" +
-                "sDA3oDWgM4YxaHR0cDovL2NybC50ZXN0NC5idXlwYXNzLm5vL2NybC9CUENsYXNz" +
-                "M1Q0Q0EzLmNybDB1oHOgcYZvbGRhcDovL2xkYXAudGVzdDQuYnV5cGFzcy5uby9k" +
-                "Yz1CdXlwYXNzLGRjPU5PLENOPUJ1eXBhc3MlMjBDbGFzcyUyMDMlMjBUZXN0NCUy" +
-                "MENBJTIwMz9jZXJ0aWZpY2F0ZVJldm9jYXRpb25MaXN0MIGKBggrBgEFBQcBAQR+" +
-                "MHwwOwYIKwYBBQUHMAGGL2h0dHA6Ly9vY3NwLnRlc3Q0LmJ1eXBhc3Mubm8vb2Nz" +
-                "cC9CUENsYXNzM1Q0Q0EzMD0GCCsGAQUFBzAChjFodHRwOi8vY3J0LnRlc3Q0LmJ1" +
-                "eXBhc3Mubm8vY3J0L0JQQ2xhc3MzVDRDQTMuY2VyMA0GCSqGSIb3DQEBCwUAA4IB" +
-                "AQCe67UOZ/VSwcH2ov1cOSaWslL7JNfqhyNZWGpfgX1c0Gh+KkO3eVkMSozpgX6M" +
-                "4eeWBWJGELMiVN1LhNaGxBU9TBMdeQ3SqK219W6DXRJ2ycBtaVwQ26V5tWKRN4Ul" +
-                "RovYYiY+nMLx9VrLOD4uoP6fm9GE5Fj0vSMMPvOEXi0NsN+8MUm3HWoBeUCLyFpe" +
-                "7/EPsS/Wud5bb0as/E2zIztRodxfNsoiXNvWaP2ZiPWFunIjK1H/8EcktEW1paiP" +
-                "d8AZek/QQoG0MKPfPIJuqH+WJU3a8J8epMDyVfaek+4+l9XOeKwVXNSOP/JSwgpO" +
-                "JNzTdaDOM+uVuk75n2191Fd7");
-    }
-}
